requirements.txt
defusedxml==0.7.1
fonttools==4.59.0
fpdf2==2.8.3
pillow==11.3.0
PyYAML==6.0.2
